requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.5
django-filter==23.2
django-fsm==2.8.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.26.4
exceptiongroup==1.1.3
filelock==3.12.3
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
pyproject-api==1.6.1
pytest==7.4.1
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.2
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.4
